feat(datepicker): add the ability to disable specific days of the week to the datepicker
Change the date picker to add the ability to choose to disable specific
days of the week.
Adds an option, daysOfWeekDisabled, that is a string with the default
value of '' (empty string).
To disable days of the week, append the appropriate value for the day
of the week that is
desired to be disabled, with 0=Sun, 1=Mon, etc. For example, to disable
Sat and Sun, add the
attribute days-of-week-disabled=06 to the element when using the
datepicker directive.