fix(timepicker): fix empty value parsing and validation
In AngularJS 1.3, if the $parsers callback returns an 'undefined' value, it causes the ngModelController to invalidate the ngModel and not run the validators. When the input is an empty string, the $parsers callback was changed to return 'null' instead of 'undefined', so we consider an empty string a valid value and allow for other validators to run and check the value.

This means a BREAKING CHANGE from previous versions, because the model value for an empty input is no longer 'undefined' but is now 'null'.