feat(modal): set incremental z-index value for modals and asides to allow for stacking
The number of active modals and asides are tracked in order to set a `z-index` value that will allow those elements to be on top of each other.
Currently, if two asides are shown at the same time, clicking the backdrop will hide the aside that is below the one above.  Also, the backdrop of the second aside is below the first aside.