feat(callbacks): add function callbacks for events via attributes on directives
To perform an action on an event that a component will trigger (e.g. show a modal)
you have to listen to the event on the scope.
Now, you can attach the function to be called via attributes (`bs-on-*`).  This is used
for all events that are also emitted through the scope.